move middleware stack into a place more similar to other parts of metrician: lib/metrician/middleware
